Improved: Added WidgetUtils file having commonly used widgets for showing information
