+++ src/pages/login/login.ts
@@ -1,10 +1,8 @@
 import { Component } from '@angular/core';
-import {
-  IonicPage, NavController, NavParams, LoadingController, Loading, ToastController,
-  AlertController
-} from 'ionic-angular';
+import { IonicPage, NavController, NavParams, Loading } from 'ionic-angular';
 import { AuthProvider } from '../../providers/auth/auth';
-import {LoginModel} from "../../models/login";
+import { LoginModel } from "../../models/login";
+import { Dialog } from "../../shared/widget.util";
 
 @IonicPage()
 @Component({
@@ -15,80 +13,34 @@ export class LoginPage {
   loader: Loading;
   companyLogo: string;
 
-  constructor(public navCtrl: NavController, public navParams: NavParams, public auth: AuthProvider, public loadingCtrl: LoadingController, public toastCtrl: ToastController, public alertCtrl: AlertController) {
+  constructor(private navCtrl: NavController, private navParams: NavParams, private auth: AuthProvider, private dialog: Dialog) {
     this.companyLogo =  '../../assets/imgs/hc.png';
   }
 
   getAuthToken(username, password){
     if(localStorage.getItem('baseUrl') == null) {
-      this.showToast('Please Enter Tenant URL before logging in');
+      this.dialog.showToast('Please Enter Tenant URL before logging in');
     }
     else {
-      this.showLoading();
+      this.dialog.showLoading();
       this.auth.doLogin(username, password).subscribe(
         (data: LoginModel) => {
           localStorage.setItem('token', data.token);
-          this.hideLoading();
+          this.dialog.hideLoading();
           this.navCtrl.setRoot('HomePage', {data: data});
         },
         (err) => {
           if (this.loader)
-            this.hideLoading();
-          this.showToast(err.statusText);
+            this.dialog.hideLoading();
+          this.dialog.showToast(err.statusText);
           console.log(err);
         }
       );
     }
   }
 
-  showLoading() {
-    this.loader = this.loadingCtrl.create({
-      content: "Logging In..."
-    });
-    this.loader.present();
-  }
-
-  hideLoading() {
-    this.loader.dismiss();
-  }
-
-  showToast(message) {
-    let toast = this.toastCtrl.create({
-      message: message,
-      duration: 3000
-    });
-    toast.present();
-  }
-
   showPrompt() {
-    let prompt = this.alertCtrl.create({
-      title: 'Tenant URL',
-      message: "Enter the name of instance. Do not enter complete URL, just enter the name of instance",
-      inputs: [
-        {
-          name: 'url',
-          placeholder: 'Tenant URL',
-          value : localStorage.getItem('baseUrl')
-        },
-      ],
-      buttons: [
-        {
-          text: 'Cancel',
-          handler: data => {
-            console.log('Cancel clicked');
-          }
-        },
-        {
-          text: 'Save',
-          handler: data => {
-            localStorage.removeItem('baseUrl');
-            localStorage.setItem('baseUrl', data.url);
-            console.log('Saved clicked');
-          }
-        }
-      ]
-    });
-    prompt.present();
-
+    this.dialog.showPrompt();
   }
+
 }
